Run diagnosis function

This function runs the diagnosis analysis for replication studies. It returns a decomposition of the discrepancy between a pair of original and replication studies into pieces attributable to covariate shift, mediation shift, and residual factors. It also provides post-selective confidence intervals for these pieces that adjust for potential publication bias.

run_diagnosis(
  org_df,
  rep_df,
  analysis_formula,
  treatment_name,
  id_name_1 = NULL,
  id_name_2 = NULL,
  covariate_formula = NULL,
  covariates = NULL,
  mediation_formula = NULL,
  mediators = NULL,
  alpha = 0.1,
  verbose = TRUE,
  if_selective = TRUE,
  pub_pvalue_threshold = 0.05
)

Arguments

org_df

The data frame for the original study

rep_df

The data frame for the replication study

analysis_formula

The analysis formula for the regression, such as "outcome ~ treatment"

treatment_name

String for the column name of the treatment indicator (=1 means treatment)

id_name_1

Optional, string for the column name of the cluster id in the original study (specify if the study is clustered)

id_name_2

Optional, string for the column name of the cluster id in the original study (specify if the study is clustered)

covariate_formula

Optional, formula for covariates to balance, such as "~ age + gender" which means the covariates to balance are "age" and "gender"

covariates

Optional, vector for column names or column number of covariates to balance. At least one of covariates and covariate_formula must be specified. If both are specified, covariate_formula will be taken

mediation_formula

Optional, formula for mediators to balance, which should also include covariates, such as "~ (age + gender + mediator) * treatment", which means "mediator" is viewed as a mediator to balance

mediators

Optional, vector for column names or column number of mediators to balance. At least one of mediation_formula and mediators must be specified. If both are specified, mediation_formula is taken

alpha

Optional, default at 0.1, coverage level for the confidence intervals is 1-alpha

verbose

Optional, default at TRUE, whether to display summary tables and decomposition plots

if_selective

Optional, default at TRUE, whether or not to run selective inference that adjusts for publication bias

pub_pvalue_threshold

Optional, default at 0.05, the threshold for publication bias, i.e., the selective inference CIs are valid conditonal on that the original study p-value is smaller than pub_pvalue_threshold

Value

A list containing the summary table, decomposition plot, summary table for selective inference, decomposition plot for selective inference, and error message for selective inference in the case of infinite confidence intervals